requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
openai==1.3.7
qdrant-client==1.7.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings
python-dotenv==1.0.0
pypdf2==3.0.1
pandas==2.1.4